+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Tx.hs
@@ -10,6 +10,7 @@
 {-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE NamedFieldPuns #-}
+{-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE RecordWildCards #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE StandaloneDeriving #-}
@@ -101,12 +102,15 @@ import Cardano.Ledger.BaseTypes (ProtVer, integralToBounded)
 import Cardano.Ledger.Binary (
   Annotator,
   DecCBOR (..),
+  Decoder,
   EncCBOR (encCBOR),
   Encoding,
   ToCBOR (..),
   decodeNullStrictMaybe,
   encodeListLen,
   encodeNullStrictMaybe,
+  ifDecoderVersionAtLeast,
+  natVersion,
   serialize,
   serialize',
  )
@@ -449,6 +453,21 @@ instance
   where
   toCBOR = toEraCBOR @era
 
+decodeAlonzoTxPv12 ::
+  forall era s.
+  ( DecCBOR (Annotator (TxBody TopTx era))
+  , DecCBOR (Annotator (TxWits era))
+  , DecCBOR (Annotator (TxAuxData era))
+  ) =>
+  Decoder s (Annotator (AlonzoTx TopTx era))
+decodeAlonzoTxPv12 = decodeRecordNamed "AlonzoTx" (const 4) $ do
+  body <- decCBOR
+  wits <- decCBOR
+  isValid <- decCBOR
+  auxData <- decodeNullStrictMaybe decCBOR
+  pure $ AlonzoTx <$> body <*> wits <*> pure isValid <*> sequence auxData
+{-# INLINE decodeAlonzoTxPv12 #-}
+
 instance
   ( Typeable l
   , Era era
@@ -464,12 +483,13 @@ instance
   decCBOR =
     withSTxTopLevelM @l @era $ \case
       STopTxOnly ->
-        decode $
-          Ann (RecD AlonzoTx)
-            <*! From
-            <*! From
-            <*! Ann From
-            <*! D (sequence <$> decodeNullStrictMaybe decCBOR)
+        ifDecoderVersionAtLeast (natVersion @12) decodeAlonzoTxPv12 $
+          decode $
+            Ann (RecD AlonzoTx)
+              <*! From
+              <*! From
+              <*! Ann From
+              <*! D (sequence <$> decodeNullStrictMaybe decCBOR)
   {-# INLINE decCBOR #-}
 
 data AlonzoStAnnTx l era where