+++ eras/alonzo/impl/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.16.0.0
 
+* Export `alonzoPlutusScriptDecoder` from `Cardano.Ledger.Alonzo.TxWits`
+* Add `generate-cbor` executable
 * Rename the `Rewarding` verb to `Withdrawing` and deprecate the old names:
   - `AlonzoRewarding` -> `AlonzoWithdrawing`
   - `RewardingPurpose` -> `WithdrawingPurpose`